Bug 1330890 - Part 1: Spoofing the time zone as UTC when fingerprinting resistance is enabled (adopt from Tor #16622). r?ehsan
This patch makes the nsAppRunner to set environment variable 'TZ' to 'UTC' when
'privacy.resistFingerprinting' is true. By doing so, Firefox will use UTC as its
local time zone instead of the default local time zone. This prevents a browser
be fingerprinted through the local time zone.

This patch adds a new service for fingerprinting resistance, which is called
nsRFPService. This service will be put in /toolkit/components/resistfingerprinting.
This service is responsible for observing the change of pref 'privacy.resistfingerprinting'
and doing underlying jobs. This service changes environment value 'TZ' and
calls nsJSUtils::ResetTimeZone() in response to the change of the pref. This
service will be started and initiated during the first creation of nsGlobalWindow.

This service is only a nsIObserver for now. In the future, however, it will be
responsible for more fingerprinting resistance jobs, like changing prefs after
'privacy.resistfingerprinting' is changed.

MozReview-Commit-ID: 8V47ZATgrKE