Bug 1407366 - Part 2: Rename the LanguagePrompt.jsm to RFPHelper.jsm and changing the place of doing uninitialization.
This patch changes the name of LanguagePrompt.jsm to RFPHelper.jsm for
making it more general. The RFPHelper is going to not only be
responsible for the language prompt but also for the window
letterboxing.

In addition, this patch also changes the place where do the
uninitialization of the RFPHelper. The RFPHelper won't do anything if we
open a new window and then close it because of RFPHelper will be
uninited during closing and there is no way to bring it back. So, we
move uninit into the nsBrowserGlue._onQuitApplicationGranted(), which
will be called during closing the application.