Bug 1407366 - Part 3: Implementing the window letterboxing for fingerprinting resistance.
This patch implements the window letterboxing when fingerprinting
resistance is enabled. The implementation is based on adding margins
around the browser element to make the content viewport size be always
quantized. Whenever the browser element is resized, the RFPHelper will
add margins around it. But it won't add any margins for the browser
which loads a content with the system principal. In addition, it will
also make the margin area black when a window enters the fullscreen
mode.