Bug 1586707 - Making the test 'browser/components/originattributes/test/browser/browser_cache.js' working in Fission. r?baku
This patch also adds a workaround for avoid a timing issue. The timing
issue exists if we close and rapidly open again of a given origin which
is using a shared worker with container enabled. The showdown of the
content process would be delayed in this case, so the content process
will be shutdown during opening the tab, which would lead to an
unexpected behavior.

The workaround makes the test script to wait the idle task happens after
closing the tab, which ensures the shutdown process has been finished
and can safely open the same origin again.